requirements.txt
affine==2.4.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cligj==0.7.2
Django==6.0.4
djangorestframework==3.17.1
geopandas==1.1.3
idna==3.13
numpy==2.4.4
packaging==26.2
pandas==3.0.2
psycopg2-binary==2.9.12
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
rasterio==1.5.0
requests==2.33.1
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
